+++ firmware/common/si5351c_drv.h
@@ -29,6 +29,7 @@ extern "C"
 #endif
 
 #include <stdint.h>
+#include <stddef.h>
 
 typedef struct {
 	uint8_t i2c_address;
@@ -36,7 +37,7 @@ typedef struct {
 
 void si5351c_write_single(si5351c_driver_t* const drv, uint8_t reg, uint8_t val);
 uint8_t si5351c_read_single(si5351c_driver_t* const drv, uint8_t reg);
-void si5351c_write(si5351c_driver_t* const drv, const uint8_t* const data, const uint_fast8_t data_count);
+void si5351c_write(si5351c_driver_t* const drv, const uint8_t* const data, const size_t data_count);
 
 #ifdef __cplusplus
 }
